두 종류의 FailPageGenerator들을 통합
- 통합 이전: LoginFailPageGenerator, RegistrationFailPageGenerator
- 통합 이후: FailPageGenerator
