+++ week07/loginRegister/황인우/app/src/main/java/utils/RegistrationFormChecker.java
@@ -5,7 +5,8 @@ import repositories.AccountRepository;
 import java.util.Map;
 
 public class RegistrationFormChecker {
-  public static final String INSUFFICIENT = "INSUFFICIENT";
+  public static final String INSUFFICIENT_REGISTRATION_INPUTS
+      = "INSUFFICIENT_REGISTRATION_INPUTS";
   public static final String ALREADY_EXISTING_ID = "ALREADY_EXISTING_ID";
   public static final String NOT_EQUAL_TO_PASSWORD_CHECK = "NOT_EQUAL_TO_PASSWORD_CHECK";
   public static final String ACCEPTED = "ACCEPTED";
@@ -22,7 +23,7 @@ public class RegistrationFormChecker {
         || !formData.containsKey("password")
         || !formData.containsKey("password-check")
         || !formData.containsKey("email")) {
-      return RegistrationFormChecker.INSUFFICIENT;
+      return RegistrationFormChecker.INSUFFICIENT_REGISTRATION_INPUTS;
     }
 
     if (accountRepository.findAccount(formData.get("id")) != null) {
